requirements.txt
absl-py==2.3.0
asgiref==3.10.0
astunparse==1.6.3
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
cryptography==46.0.3
curl_cffi==0.11.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flatbuffers==25.2.10
frozendict==2.4.6
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
joblib==1.5.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
protobuf==3.19.6
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
soupsieve==2.7
sqlparse==0.5.3
ta==0.11.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.63
zipp==3.23.0
